# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateMessageBody": {
        "additionalProperties": false,
        "description": "Request body for appending a message and retrieval episode.",
        "properties": {
          "content": {
            "minLength": 1,
            "title": "Content",
            "type": "string"
          },
          "role": {
            "enum": [
              "system",
              "user",
              "assistant"
            ],
            "title": "Role",
            "type": "string"
          },
          "token_count": {
            "minimum": 0.0,
            "title": "Token Count",
            "type": "integer"
          }
        },
        "required": [
          "role",
          "content",
          "token_count"
        ],
        "title": "CreateMessageBody",
        "type": "object"
      },
...
      "CreatedMessageResponse": {
        "additionalProperties": false,
        "description": "HTTP representation of a message created with its retrieval episode.",
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "conversation_id": {
            "format": "uuid",
            "title": "Conversation Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "episode_id": {
            "format": "uuid",
            "title": "Episode Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "position": {
            "title": "Position",
            "type": "integer"
          },
          "role": {
            "enum": [
              "system",
              "user",
              "assistant"
            ],
            "title": "Role",
            "type": "string"
          },
          "token_count": {
            "title": "Token Count",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "conversation_id",
          "position",
          "role",
          "content",
          "token_count",
          "created_at",
          "episode_id"
        ],
        "title": "CreatedMessageResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MessageResponse": {
        "additionalProperties": false,
        "description": "HTTP representation of an immutable conversation message.",
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "conversation_id": {
            "format": "uuid",
            "title": "Conversation Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "position": {
            "title": "Position",
            "type": "integer"
          },
          "role": {
            "enum": [
              "system",
              "user",
              "assistant"
            ],
            "title": "Role",
            "type": "string"
          },
          "token_count": {
            "title": "Token Count",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "conversation_id",
          "position",
          "role",
          "content",
          "token_count",
          "created_at"
        ],
        "title": "MessageResponse",
        "type": "object"
      },
      "MessageRetrievalsResponse": {
        "additionalProperties": false,
        "description": "HTTP representation of recorded retrieval provenance for one assistant message.",
        "properties": {
          "assistant_message_id": {
            "format": "uuid",
            "title": "Assistant Message Id",
            "type": "string"
          },
          "retrievals": {
            "items": {
              "$ref": "#/components/schemas/RetrievalEntry"
            },
            "title": "Retrievals",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "assistant_message_id",
          "total",
          "retrievals"
        ],
        "title": "MessageRetrievalsResponse",
        "type": "object"
      },
      "RetrievalEntry": {
        "additionalProperties": false,
        "description": "HTTP representation of one recorded retrieval provenance row.",
        "properties": {
          "access_score": {
            "title": "Access Score",
            "type": "number"
          },
          "episode": {
            "$ref": "#/components/schemas/RetrievalEpisode"
          },
          "frequency_score": {
            "title": "Frequency Score",
            "type": "number"
          },
          "importance_score": {
            "title": "Importance Score",
            "type": "number"
          },
          "query": {
            "$ref": "#/components/schemas/RetrievalQuery"
          },
          "rank": {
            "title": "Rank",
            "type": "integer"
          },
          "recency_score": {
            "title": "Recency Score",
            "type": "number"
          },
          "retrieved_at": {
            "format": "date-time",
            "title": "Retrieved At",
            "type": "string"
          },
          "score": {
            "title": "Score",
            "type": "number"
          },
          "scoring_version": {
            "title": "Scoring Version",
            "type": "string"
          },
          "similarity": {
            "title": "Similarity",
            "type": "number"
          }
        },
        "required": [
          "rank",
          "similarity",
          "score",
          "recency_score",
          "access_score",
          "frequency_score",
          "importance_score",
          "scoring_version",
          "retrieved_at",
          "query",
          "episode"
        ],
        "title": "RetrievalEntry",
        "type": "object"
      },
      "RetrievalEpisode": {
        "additionalProperties": false,
        "description": "HTTP representation of the episode source used by a retrieval result.",
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "enum": [
              "message",
              "summary"
            ],
            "title": "Kind",
            "type": "string"
          },
          "source_conversation_id": {
            "format": "uuid",
            "title": "Source Conversation Id",
            "type": "string"
          },
          "source_conversation_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Conversation Title"
          },
          "source_scope_id": {
            "format": "uuid",
            "title": "Source Scope Id",
            "type": "string"
          },
          "source_scope_name": {
            "title": "Source Scope Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "content",
          "source_conversation_id",
          "source_conversation_title",
          "source_scope_id",
          "source_scope_name"
        ],
        "title": "RetrievalEpisode",
        "type": "object"
      },
      "RetrievalQuery": {
        "additionalProperties": false,
        "description": "HTTP representation of the query message that produced retrieval provenance.",
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "message_id": {
            "format": "uuid",
            "title": "Message Id",
            "type": "string"
          }
        },
        "required": [
          "message_id",
          "content"
        ],
        "title": "RetrievalQuery",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Smriti Local API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/conversations/{conversation_id}/messages": {
      "get": {
        "description": "List messages for a local user's conversation.",
        "operationId": "list_messages_conversations__conversation_id__messages_get",
        "parameters": [
          {
            "in": "path",
            "name": "conversation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Conversation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MessageResponse"
                  },
                  "title": "Response List Messages Conversations  Conversation Id  Messages Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Messages",
        "tags": [
          "messages"
        ]
      },
      "post": {
        "description": "Append a message and retrieval episode for a local user's conversation.",
        "operationId": "create_message_conversations__conversation_id__messages_post",
        "parameters": [
          {
            "in": "path",
            "name": "conversation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Conversation Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMessageBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedMessageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Message",
        "tags": [
          "messages"
        ]
      }
    },
    "/conversations/{conversation_id}/messages/{message_id}/retrievals": {
      "get": {
        "description": "List recorded retrieval provenance for a local user's assistant message.",
        "operationId": "list_message_retrievals_conversations__conversation_id__messages__message_id__retrievals_get",
        "parameters": [
          {
            "in": "path",
            "name": "conversation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Conversation Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "message_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Message Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageRetrievalsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Message Retrievals",
        "tags": [
          "messages"
        ]
      }
    },
...
  }
}